requirements.txt
numpy>=1.19.4
crepe>=0.0.11
librosa>=0.8.0
einops>=0.3.0
tqdm>=4.46.0
torch>=1.7.0
torchviz
effortless_config>=0.7.0
SoundFile>=0.10.3.post1
Flask>=1.1.2
PyYAML>=5.3.1
tensorflow
tensorboard